Fix crashanalytics test - follow on from 23347 (#23369)
In https://github.com/matomo-org/matomo/pull/23347 we introduced a
change which caused a crash in the CrashAnalytics plugin, caught by an
automated test.

This change stops the error from happening.

The failed test can be found in the 5.x-dev branch by running:

ddev matomo:console tests:run-ui --plugin=CrashAnalytics --debug -- CrashAnalytics_CrashDetails

I'm not sure if this is the best fix, not being familiar with the code
area yet, but it'll be a good start at least.

Co-authored-by: James Hill <james-hill-matomo@users.noreply.github.com>